community_activity/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.7
fastapi==0.104.1
gunicorn==21.2.0
h11==0.14.0
idna==3.6
load-dotenv==0.1.0
oracledb==1.4.2
packaging==23.2
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
